ml/requirements.txt
scikit-learn==1.4.2
xgboost==2.0.3
lightgbm==4.3.0
onnx==1.16.0
onnxruntime==1.18.0
skl2onnx==0.17.0
pandas==2.2.2
numpy==1.26.4
shap==0.45.1
mlflow==2.12.2
pyyaml==6.0.1
joblib==1.4.0
scipy==1.13.0
matplotlib==3.8.4
seaborn==0.13.2
python-dotenv==1.0.1
fastapi==0.111.0
uvicorn[standard]==0.29.0
